+++ NEWS.md
@@ -1,3 +1,5 @@
 # pryr 0.1.0.9000
 
 * `address()` no longer changes `NAMED()` status of x (#24).
+
+* Use non-internal `nonS3Methods()` where needed (#38).